// lib/enums.ts
export enum EquipmentId {
    Dumbbells = 1,
    Barbell = 2,
    Kettlebell = 3,
    ResistanceBands = 4,
    PullUpBar = 5,
    Treadmill = 6,
    StationaryBike = 7,
    JumpRope = 8,
    MedicineBall = 9,
    YogaMat = 10,
    FoamRoller = 11,
    CableMachine = 12,
    RowingMachine = 13,
    Bench = 14,
    SquatRack = 15,
}

export enum ExerciseId {
    BenchPress = 1,
    OverheadPress = 2,
    PushUps = 3,
    PullUps = 4,
    BarbellRow = 5,
    Deadlift = 6,
    Squats = 7,
    Lunges = 8,
    Jog5km = 9,
    HIITSprints = 10,
    RowingMachine = 11,
    YogaFlow = 12,
    Plank = 13,
    SingleLegBalance = 14,
}